refactor: revert DS to DAO refactor(angular-http): active record has separate logic + renaming refactor: actionController does not handle model actions, moved to active record implementation.
This commit is an initial move towards cancelling TargetAdapterMetadataStore.
Instead, an ActionController will be created for every Target - Adapter pair.
This will allow invoking DAO commands with multiple adapters on the same target.
ActiveRecord pattern remains limited to the active adapter on the target, for obvious reasons.

All other logic handled in TargetAdapterMetadataStore will be moved to the TargetMetadata.