refactor: remove TargetAdapterMetadataStore refactor: change old active-record TDModel and ActiceRecordCollection to TDMModel and TDMCollection. refactor: attached resource-control no longer use events, base class is used instead (TDMModelBase)
This commit remove the use of TargetAdapterMetadataStore.
Instead, an ActionController will be created for every Target - Adapter pair.
This allows invoking DAO commands with multiple adapters on the same target.
ActiveRecord pattern remains limited to the active adapter on the target, for obvious reasons.
The build() is smaller now, it does not include target actions, this is done in the active-record plugin.

All other logic handled in TargetAdapterMetadataStore (i.e. build) is moved to TargetMetadata.