refactor(core): Majority of the code move to internal extension [@tdm/core/tdm], public [@tdm/core] is light
Moving the most of the code of core into an internal plugin "hides" the mess of a lot of exports that are only for extensibility.
This is required since the old way of exporting a module from a module is not supported by the anuglar compiler which will reflect AOT for packages using core.

refactor(build): Build Flatten ESM modules with angular support.

feat(build): support local (internal) packages

Local packages are like plugins / extension / tools that are added to the bundle of the package.
A good exapmle is a "testing" extension to support testing.
The current support make local packages as if they are native, but just internal to parent package.